X.26 lint all (#621)
* review with lint:all

* review with lint:all

* Add -Xlint:all,-options as default compiler

+++ src/test/java/net/openhft/chronicle/bytes/internal/DecimaliserFloatTest.java
@@ -75,7 +75,7 @@ class DecimaliserFloatTest extends BytesTestCommon {
             assertEquals(48847, mantissa);
             assertEquals(16, exponent);
         };
-        assertTrue(new MaximumPrecision(16).toDecimal((float) HARD_TO_DECIMALISE, check));
+        assertTrue(new MaximumPrecision(16).toDecimal(HARD_TO_DECIMALISE, check));
     }
 
     @Test
